Synchronizing to Wordpress launching

+++ ChupooTemplateEngine/LayoutParsers/Wordpress.cs
@@ -46,6 +46,13 @@ namespace ChupooTemplateEngine.LayoutParsers
 
             Directories.Current = Directories.Layout;
             layout_content = File.ReadAllText(path);
+            
+            LibParser lp = new LibParser();
+            layout_content = lp.Parse(dest, layout_content);
+
+            ModuleParser mp = new ModuleParser();
+            layout_content = mp.Parse(layout_content);
+
             layout_content = RenderPartialLayout(layout_content);
             layout_content = RenderLayoutComponent(cfg_layout_name, layout_content);
 
@@ -61,7 +68,10 @@ namespace ChupooTemplateEngine.LayoutParsers
             if (Directory.Exists(Directories.Public + p_dir))
             {
                 CopyResources();
-                
+
+                layout_content = ReplaceLinkUrlText(layout_content, asset_level);
+                layout_content = ReplaceAssetUrlText(layout_content, asset_level, Directories.Layout);
+
                 if (!partial_files.ContainsKey("header"))
                 {
                     layout_content = PasteStyles(layout_content);
@@ -70,9 +80,6 @@ namespace ChupooTemplateEngine.LayoutParsers
                 {
                     layout_content = PasteScripts(layout_content);
                 }
-
-                layout_content = ReplaceLinkUrlText(layout_content, asset_level);
-                layout_content = ReplaceAssetUrlText(layout_content, asset_level, Directories.Layout);
 
                 // Make <layout>.php file
                 string p_file;
@@ -84,7 +91,6 @@ namespace ChupooTemplateEngine.LayoutParsers
                 CreatePartialFile("footer", asset_level);
 
                 partial_files.Clear();
-                ClearAll();
             }
             else
             {
@@ -97,7 +103,15 @@ namespace ChupooTemplateEngine.LayoutParsers
             if (partial_files.ContainsKey(partialName))
             {
                 string part_content = File.ReadAllText(partial_files[partialName] + "");
+
+                LibParser lp = new LibParser();
+                part_content = lp.Parse(partialName, part_content);
+
+                ModuleParser mp = new ModuleParser();
+                part_content = mp.Parse(part_content);
+
                 part_content = RenderPartialLayout(part_content);
+                part_content = ReplaceLinkUrlText(part_content, asset_level);
                 if (partialName == "header")
                 {
                     part_content = PasteStyles(part_content);
@@ -106,7 +120,6 @@ namespace ChupooTemplateEngine.LayoutParsers
                 {
                     part_content = PasteScripts(part_content);
                 }
-                part_content = ReplaceLinkUrlText(part_content, asset_level);
                 part_content = ReplaceAssetUrlText(part_content, asset_level, Directories.Layout);
                 string p_file = Directories.Public + partialName + ".php";
                 File.WriteAllText(p_file, part_content);
@@ -125,7 +138,7 @@ namespace ChupooTemplateEngine.LayoutParsers
                 {
                     string name = match.Groups[1].Value;
                     string layout_name = name;
-                    string layout_file = Directories.Layout + layout_name + ".html";
+                    string layout_file = Directories.Layout + "_" + layout_name + ".html";
 
                     if (name == "header" || name == "footer")
                     {
@@ -139,6 +152,13 @@ namespace ChupooTemplateEngine.LayoutParsers
                         if (File.Exists(layout_file))
                         {
                             string part_content = File.ReadAllText(layout_file);
+
+                            LibParser lp = new LibParser();
+                            part_content = lp.Parse(name, part_content);
+
+                            ModuleParser mp = new ModuleParser();
+                            part_content = mp.Parse(part_content);
+
                             part_content = RenderLayoutComponent(layout_name, part_content);
                             content = SubsituteString(content, match.Index + newLength, match.Length, part_content);
                             newLength += part_content.Length - match.Length;
@@ -167,34 +187,39 @@ namespace ChupooTemplateEngine.LayoutParsers
         private void CopyResources()
         {
             string dst = Directories.Public + "style.css";
-            string src = Directories.Project + "\\launch\\wordpress\\style.css";
-            File.Copy(src, dst);
+            string src = Directories.Dev + "\\launch\\wordpress\\style.css";
+            if (!File.Exists(dst))
+            {
+                File.Copy(src, dst);
+            }
         }
 
-        private void CreateFunctionsFile()
+        public static void CreateFunctionsFile()
         {
             // Make functions.php file
             string p_file = Directories.Public + "functions.php";
             Hashtable data = new Hashtable();
-            data["ThemeName"] = "swad";
+            data["ThemeName"] = "ail";
 
             List<Hashtable> JsFiles = new List<Hashtable>();
             int i = 0;
             foreach(string file in script_file_list)
             {
                 Hashtable _file = new Hashtable();
-                _file["Id"] = "script" + i;
-                _file["Path"] = file;
+                _file["Id"] = "c.asset" + i;
+                _file["Path"] = "/" + file;
                 JsFiles.Add(_file);
                 i++;
             }
+            data["Js"] = JsFiles;
 
+            i = 0;
             List<Hashtable> CssFiles = new List<Hashtable>();
             foreach (string file in style_file_list)
             {
                 Hashtable _file = new Hashtable();
-                _file["Id"] = "style" + i;
-                _file["Path"] = file;
+                _file["Id"] = "c.asset" + i;
+                _file["Path"] = "/" + file;
                 CssFiles.Add(_file);
                 i++;
             }
